test: cover generated action metadata snapshot Signed-off-by: Gregor Zeitlinger <gregor.zeitlinger@grafana.com>
