+++ src/linters/renovate_deps/mod.rs
@@ -3,6 +3,7 @@ use std::collections::HashSet;
 use std::path::{Path, PathBuf};
 use std::process::Stdio;
 
+use self::action_validation::validate_lookup_action_warnings;
 use self::install_patch::configure_extract_workaround_env;
 use self::manager_patterns::changed_matches_manager_file_patterns;
 use self::mise_normalize::patch_semver_equivalent_mise_values;
@@ -11,9 +12,7 @@ use self::rules::{
     incomplete_meta_for_rules, trim_snapshot_meta, validate_extract_version_consistency,
     validate_rule_coverage,
 };
-use self::snapshot::{
-    Snapshot, canonical_manager_name, extract_deps, read_snapshot, unified_diff, write_snapshot,
-};
+use self::snapshot::{Snapshot, extract_deps, read_snapshot, unified_diff, write_snapshot};
 use crate::config::RenovateDepsConfig;
 use crate::files::FileList;
 use crate::linters::LinterOutput;
@@ -23,6 +22,7 @@ use crate::registry::{
     NativeRunContext, NativeRunFuture, PreparedNativeCheck,
 };
 
+mod action_validation;
 mod install_patch;
 mod manager_patterns;
 mod mise_normalize;
@@ -675,75 +675,6 @@ async fn generate_snapshot(
         tokio::time::sleep(std::time::Duration::from_secs(3)).await;
     }
     Ok(generated)
-}
-
-/// Renovate normally exits successfully even when a digest lookup cannot be
-/// completed. A missing digest for a GitHub Action is actionable when
-/// Renovate reports its deterministic error, but `no-result`, authentication,
-/// and network failures are inconclusive and must not make lint flaky.
-fn validate_lookup_action_warnings(
-    log_bytes: &[u8],
-    exclude_managers: &[String],
-) -> anyhow::Result<()> {
-    const DETERMINISTIC: &str = "Could not determine new digest for update";
-
-    if exclude_managers
-        .iter()
-        .map(|manager| canonical_manager_name(manager))
-        .any(|manager| manager == "github-actions")
-    {
-        return Ok(());
-    }
-
-    for line in std::str::from_utf8(log_bytes)?.lines() {
-        let Ok(entry) = serde_json::from_str::<serde_json::Value>(line) else {
-            continue;
-        };
-        if !entry
-            .get("msg")
-            .and_then(|value| value.as_str())
-            .is_some_and(|msg| msg == "packageFiles with updates")
-        {
-            continue;
-        }
-        let Some(package_files) = entry
-            .get("packageFiles")
-            .or_else(|| entry.get("config"))
-            .and_then(|value| value.get("github-actions"))
-            .and_then(|value| value.as_array())
-        else {
-            continue;
-        };
-        for package_file in package_files {
-            let Some(deps) = package_file.get("deps").and_then(|value| value.as_array()) else {
-                continue;
-            };
-            for dep in deps {
-                let Some(warnings) = dep.get("warnings").and_then(|value| value.as_array()) else {
-                    continue;
-                };
-                for warning in warnings {
-                    let Some(message) = warning.get("message").and_then(|value| value.as_str())
-                    else {
-                        continue;
-                    };
-                    if !message.starts_with(DETERMINISTIC) {
-                        continue;
-                    }
-                    let dep_name = dep
-                        .get("depName")
-                        .or_else(|| dep.get("packageName"))
-                        .and_then(|value| value.as_str())
-                        .unwrap_or("unknown dependency");
-                    anyhow::bail!(
-                        "Renovate reported an invalid GitHub Action ref for {dep_name}: {message}"
-                    );
-                }
-            }
-        }
-    }
-
-    Ok(())
 }
 
 /// Runs `renovate --platform=local` and returns the combined stdout+stderr log bytes.
