+++ tests/cases/renovate-deps/action-metadata-populated/files/.github/renovate.json5
@@ -1,0 +1,3 @@
+{
+  extends: ["config:recommended"]
+}
