+++ tests/cases/renovate-deps/action-metadata-populated/files/mise.toml
@@ -1,0 +1,2 @@
+[tools]
+"npm:renovate" = "43.150.0"
